requirements.txt
disputils==0.1.3
aiohttp~=3.7.4.post0
asyncpg~=0.22.0
discord.py==1.6.0
